+++ apis/conversion/src/__tests__/array-util.inspect.js
@@ -7,17 +7,17 @@ describe('array util', () => {
       const arr2 = [2, 3];
       const arr3 = [0, 3];
       const arr4 = [0, 3, 5];
-      expect(arrayUtil.isOrderedSubset(arr1, arr2)).to.be.true;
-      expect(arrayUtil.isOrderedSubset(arr1, arr3)).to.be.true;
-      expect(arrayUtil.isOrderedSubset(arr1, arr4)).to.be.true;
+      expect(arrayUtil.isOrderedSubset(arr1, arr2)).toBe(true);
+      expect(arrayUtil.isOrderedSubset(arr1, arr3)).toBe(true);
+      expect(arrayUtil.isOrderedSubset(arr1, arr4)).toBe(true);
     });
 
     it('arrays is not subset', () => {
       const arr1 = [0, 1, 2, 3, 4, 5];
       const arr2 = [3, 2];
       const arr3 = [1, 3, 2];
-      expect(arrayUtil.isOrderedSubset(arr1, arr2)).to.be.false;
-      expect(arrayUtil.isOrderedSubset(arr1, arr3)).to.be.false;
+      expect(arrayUtil.isOrderedSubset(arr1, arr2)).toBe(false);
+      expect(arrayUtil.isOrderedSubset(arr1, arr3)).toBe(false);
     });
   });
 
@@ -27,8 +27,8 @@ describe('array util', () => {
       const arr2 = [1, 0, 2];
       arrayUtil.indexAdded(arr1, 0);
       arrayUtil.indexAdded(arr2, 0);
-      expect(arr1).to.deep.equal([1, 2, 3, 0]);
-      expect(arr2).to.deep.equal([2, 1, 3, 0]);
+      expect(arr1).toEqual([1, 2, 3, 0]);
+      expect(arr2).toEqual([2, 1, 3, 0]);
     });
 
     it('should add index 1 correctly', () => {
@@ -36,8 +36,8 @@ describe('array util', () => {
       const arr2 = [1, 0, 2];
       arrayUtil.indexAdded(arr1, 1);
       arrayUtil.indexAdded(arr2, 1);
-      expect(arr1).to.deep.equal([0, 2, 3, 1]);
-      expect(arr2).to.deep.equal([2, 0, 3, 1]);
+      expect(arr1).toEqual([0, 2, 3, 1]);
+      expect(arr2).toEqual([2, 0, 3, 1]);
     });
 
     it('should add index 2 correctly', () => {
@@ -45,8 +45,8 @@ describe('array util', () => {
       const arr2 = [1, 0, 2];
       arrayUtil.indexAdded(arr1, 2);
       arrayUtil.indexAdded(arr2, 2);
-      expect(arr1).to.deep.equal([0, 1, 3, 2]);
-      expect(arr2).to.deep.equal([1, 0, 3, 2]);
+      expect(arr1).toEqual([0, 1, 3, 2]);
+      expect(arr2).toEqual([1, 0, 3, 2]);
     });
 
     it('should add index 3 correctly', () => {
@@ -54,8 +54,8 @@ describe('array util', () => {
       const arr2 = [1, 0, 2];
       arrayUtil.indexAdded(arr1, 3);
       arrayUtil.indexAdded(arr2, 3);
-      expect(arr1).to.deep.equal([0, 1, 2, 3]);
-      expect(arr2).to.deep.equal([1, 0, 2, 3]);
+      expect(arr1).toEqual([0, 1, 2, 3]);
+      expect(arr2).toEqual([1, 0, 2, 3]);
     });
   });
 
@@ -65,8 +65,8 @@ describe('array util', () => {
       const arr2 = [2, 3, 0, 1];
       arrayUtil.indexRemoved(arr1, 0);
       arrayUtil.indexRemoved(arr2, 0);
-      expect(arr1).to.deep.equal([0, 1, 2]);
-      expect(arr2).to.deep.equal([1, 2, 0]);
+      expect(arr1).toEqual([0, 1, 2]);
+      expect(arr2).toEqual([1, 2, 0]);
     });
 
     it('should remove index 1 correctly', () => {
@@ -74,8 +74,8 @@ describe('array util', () => {
       const arr2 = [2, 3, 0, 1];
       arrayUtil.indexRemoved(arr1, 1);
       arrayUtil.indexRemoved(arr2, 1);
-      expect(arr1).to.deep.equal([0, 1, 2]);
-      expect(arr2).to.deep.equal([1, 2, 0]);
+      expect(arr1).toEqual([0, 1, 2]);
+      expect(arr2).toEqual([1, 2, 0]);
     });
 
     it('should remove index 2 correctly', () => {
@@ -83,8 +83,8 @@ describe('array util', () => {
       const arr2 = [2, 3, 0, 1];
       arrayUtil.indexRemoved(arr1, 2);
       arrayUtil.indexRemoved(arr2, 2);
-      expect(arr1).to.deep.equal([0, 1, 2]);
-      expect(arr2).to.deep.equal([2, 0, 1]);
+      expect(arr1).toEqual([0, 1, 2]);
+      expect(arr2).toEqual([2, 0, 1]);
     });
 
     it('should remove index 3 correctly', () => {
@@ -92,8 +92,8 @@ describe('array util', () => {
       const arr2 = [2, 3, 0, 1];
       arrayUtil.indexRemoved(arr1, 3);
       arrayUtil.indexRemoved(arr2, 3);
-      expect(arr1).to.deep.equal([0, 1, 2]);
-      expect(arr2).to.deep.equal([2, 0, 1]);
+      expect(arr1).toEqual([0, 1, 2]);
+      expect(arr2).toEqual([2, 0, 1]);
     });
 
     it('should remove index 4 correctly', () => {
@@ -101,8 +101,8 @@ describe('array util', () => {
       const arr2 = [2, 3, 0, 1];
       arrayUtil.indexRemoved(arr1, 4);
       arrayUtil.indexRemoved(arr2, 4);
-      expect(arr1).to.deep.equal([1, 2, 3]);
-      expect(arr2).to.deep.equal([3, 0, 1]);
+      expect(arr1).toEqual([1, 2, 3]);
+      expect(arr2).toEqual([3, 0, 1]);
     });
   });
 });