+++ apis/nucleus/src/components/listbox/ListBoxInline.jsx
@@ -212,7 +212,7 @@ export default function ListBoxInline({ app, fieldIdentifier, stateName = '$', o
       setShowToolbar(selections.isActive());
     }
     return () => {
-      if (selections) {
+      if (selections && selections.removeListener) {
         selections.removeListener('deactivated', show);
         selections.removeListener('activated', hide);
       }
@@ -276,6 +276,7 @@ export default function ListBoxInline({ app, fieldIdentifier, stateName = '$', o
 
   return (
     <StyledGrid
+      className="listbox-container"
       container
       tabIndex={keyboard.enabled && !keyboard.active ? 0 : -1}
       direction="column"