+++ commands/serve/web/components/property-panel/DataCube.jsx
@@ -1,7 +1,8 @@
-import React, { useMemo } from 'react';
+import React, { useMemo, useState } from 'react';
 
 import hcHandler from '@nebula.js/nucleus/src/object/hc-handler';
 import loHandler from '@nebula.js/nucleus/src/object/lo-handler';
+import loContainerHandler from '@nebula.js/nucleus/src/object/lo-container-handler';
 
 import { Typography } from '@mui/material';
 
@@ -27,25 +28,39 @@ const getValue = (data, reference, defaultValue) => {
 };
 
 export default function DataCube({ setProperties, target, properties }) {
-  const createHandler = target.propertyPath.match('/qHyperCube') ? hcHandler : loHandler;
+  // Store dimensions outside of handler to support filterpanes or other containers
+  // holding list objects, which do not store dimensions in an array (like hc.qDimensions).
+  const [dimensions, setDimensions] = useState([]);
+
+  const supportsMultipleDims = target.dimensions && target.dimensions.max && target.dimensions.max() > 1;
+  const listBoxHandler = supportsMultipleDims ? loContainerHandler : loHandler;
+
+  const createHandler = target.propertyPath.match('/qHyperCube') ? hcHandler : listBoxHandler;
+  const dc = getValue(properties, target.propertyPath);
   const handler = useMemo(
     () =>
       createHandler({
         def: target,
-        dc: getValue(properties, target.propertyPath),
+        dc,
         properties,
+        dimensions,
       }),
-    [properties]
+    [dc, target, properties, dimensions]
   );
 
   const onDimensionAdded = (a) => {
-    handler.addDimension(typeof a === 'object' ? { qLibraryId: a.qId } : a);
-    setProperties(properties);
+    const dim = typeof a === 'object' ? { qLibraryId: a.qId } : a;
+    const addedDim = handler.addDimension(dim);
+    setProperties(properties).then(() => {
+      setDimensions([...dimensions, addedDim]);
+    });
   };
 
   const onDimensionRemoved = (idx) => {
     handler.removeDimension(idx);
-    setProperties(properties);
+    setProperties(properties).then(() => {
+      setDimensions([...dimensions.slice(0, idx), ...dimensions.slice(idx + 1)]);
+    });
   };
 
   const onMeasureAdded = (a) => {
@@ -64,6 +79,7 @@ export default function DataCube({ setProperties, target, properties }) {
         {target.propertyPath}
       </Typography>
       <Fields
+        key="dimensions"
         onAdded={onDimensionAdded}
         onRemoved={onDimensionRemoved}
         canAdd={handler.canAddDimension()}
@@ -73,6 +89,7 @@ export default function DataCube({ setProperties, target, properties }) {
         addLabel="Add dimension"
       />
       <Fields
+        key="measures"
         onAdded={onMeasureAdded}
         onRemoved={onMeasureRemoved}
         canAdd={handler.canAddMeasure()}